client/pyproject.toml
[project]
name = "msensor_client"
version = "0.1.0"
description = "Msensor client library"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "grpcio==1.71.0",
    "grpcio-tools==1.71.0",
    "opencv-python>=4.12.0.88",
    "viser>=1.0.16",
]

[tool.black]
line-length = 120
extend-exclude = '''
proto_gen
'''
